[FIX] 중간저장
