+++ Money-Planner/Money-Planner/Model/Goal/GoalDetail.swift
@@ -8,25 +8,10 @@
 import Foundation
 
 //goal detail 최상단을 위한 정보
-//struct GoalDetailResponse: Decodable {
-//    public let isSuccess: Bool
-//    public let message: String
-//    public let result: GoalDetail
-//}
-//
-//struct GoalDetail: Decodable {
-//    let title: String
-//    let icon: String
-//    let startDate: String
-//    let endDate: String
-//    let totalBudget: Int64
-//    let totalCost: Int64
-//}
-
 struct GoalDetailResponse: Decodable {
-    let isSuccess: Bool
-    let message: String
-    let result: GoalDetail
+    public let isSuccess: Bool
+    public let message: String
+    public let result: GoalDetail
 }
 
 struct GoalDetail: Decodable {
@@ -36,41 +21,33 @@ struct GoalDetail: Decodable {
     let endDate: String
     let totalBudget: Int64
     let totalCost: Int64
-
-    enum CodingKeys: String, CodingKey {
-        case title
-        case icon
-        case startDate = "startDate" // 예시: JSON 키가 snake_case인 경우
-        case endDate = "endDate"     // 예시: JSON 키가 snake_case인 경우
-        case totalBudget = "totalBudget" // 예시: JSON 키가 snake_case인 경우
-        case totalCost = "totalCost"     // 예시: JSON 키가 snake_case인 경우
-    }
 }
 
-
-//goaldetails - expense
-struct GoalExpenseResponse: Codable {
-    let isSuccess: Bool
-    let message: String
-    let result: GoalExpenseResult
+//expense
+struct ExpenseDetail: Decodable {
+    let expenseId: Int64
+    let title: String
+    let cost: Int64
+    let categoryIcon: String
 }
 
-struct GoalExpenseResult: Codable {
-    let dailyExpenseList: [DailyExpense]
-    let hasNext: Bool
-}
-
-struct DailyExpense: Codable {
+// DailyExpenseList 항목에 대한 구조체
+struct DailyExpense: Decodable {
     let date: String
-    let dailyTotalCost: Int
+    let dailyTotalCost: Int64
     let expenseDetailList: [ExpenseDetail]
 }
 
-struct ExpenseDetail: Codable {
-    let expenseId: Int
-    let title: String
-    let cost: Int
-    let categoryIcon: String
+// 전체 응답에 대한 구조체
+struct WeeklyExpenseResponse: Decodable {
+    let isSuccess: Bool
+    let message: String
+    let result: WeeklyExpenseResult
+}
+
+struct WeeklyExpenseResult: Decodable {
+    let dailyExpenseList: [DailyExpense]
+    let hasNext: Bool
 }
 
 
@@ -82,22 +59,22 @@ struct GoalReportResponse: Codable {
 }
 
 struct GoalReportResult: Codable {
-    let zeroDayCount: Int
+    let zeroDayCount: Int64
     let categoryTotalCosts: [CategoryTotalCost]
     let categoryGoalReports: [CategoryGoalReport]
 }
 
 struct CategoryTotalCost: Codable {
     let categoryName: String
-    let totalCost: Int
+    let totalCost: Int64
 }
 
 struct CategoryGoalReport: Codable {
     let categoryName: String
     let categoryIcon: String
-    let categoryBudget: Int
-    let totalCost: Int
-    let avgCost: Int
-    let maxCost: Int
-    let expenseCount: Int
+    let categoryBudget: Int64
+    let totalCost: Int64
+    let avgCost: Int64
+    let maxCost: Int64
+    let expenseCount: Int64
 }
